requirements.txt
django
djangorestframework
requests
pyyaml
